move SyncPeer to separate file
